refactor(storage): split structured mutation modules (#2196)
* refactor(storage): split structured mutation modules

* chore: retrigger review for mutation modules

* chore: retrigger review for mutation modules

---------
